requirements.txt
streamlit
transformers==4.52.2
torch
torchvision
pillow
accelerate